tb_ms_prediction/requirements.txt
tensorflow>=2.10.0
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
opencv-python>=4.6.0
Pillow>=9.0.0
streamlit>=1.20.0
fastapi>=0.95.0
uvicorn>=0.20.0
python-multipart>=0.0.6
kaggle>=1.5.12
tqdm>=4.64.0
plotly>=5.11.0
grad-cam>=1.4.6
pytest>=7.2.0
requests>=2.28.0
python-jose>=3.3.0
